Merge branch 'GH-6'
This adds the feature to dynamically identify binary libraries so they can be collected during the packaging process, obviating the need for them to be stored in the repository FOR OSX ONLY. The Windows portion turns out to be more difficult to do than I want to deal with right now. 

Merging this into master to work on GH-5. 

Closes GH-6. GH-7 opened for the Windows support.